refactor: rename two components
change documentation-card to documentation-section-card and dropdown-card to documentation-card
