+++ src/components/documentation-section/index.tsx
@@ -1,6 +1,6 @@
 import { Box, Flex, Text } from '@vtex/brand-ui'
 
-import DocumentationCard from '../documentation-card'
+import DocumentationSectionCard from '../documentation-section-card'
 
 import { messages } from 'utils/constants'
 import { documentationData as cards } from 'utils/constants'
@@ -15,7 +15,7 @@ const DocumentationSection = () => {
       </Text>
       <Flex sx={styles.cardsContainer} data-cy="documentation-card-list">
         {cards.map((card) => (
-          <DocumentationCard key={card.title} {...card} />
+          <DocumentationSectionCard key={card.title} {...card} />
         ))}
       </Flex>
     </Box>
